requirements.txt
langchain
wikipedia
langchain-community
langchain-openai
langchain-anthropic
python-dotenv
pydantic
fastapi
uvicorn
google-generativeai
python-dotenv
requests
aiofiles